py-api-test/requirements.txt
python-dotenv
prefect==2.8.1
psycopg2-binary

# API
fastapi
pytz
uvicorn
# tests
httpx
pytest-asyncio